requirements_ours.txt
accelerate
diffusers
# torch==2.5.1
# torchvision>=0.19.0
matplotlib
tensorboard
scikit-image
timm
transformers
deepspeed
open_clip_torch
pandas
opencv-python
gpustat
tqdm
ipdb
kornia
einops
trimesh
omegaconf
lpips
easydict
ninja
scipy
av
scikit-learn
pytorch-lightning
roma
imagesize
moviepy
plyfile
carvekit
pillow_heif
wandb
torchmetrics[image]
webdataset
xformers

# val
seaborn

# pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu126
# pip install git+https://github.com/facebookresearch/pytorch3d.git
